moving the remaining engines unit tests over to pytest.
